moved the utility to TableConfigUtils
