+++ pinot-segment-spi/src/main/java/org/apache/pinot/segment/spi/index/FieldIndexConfigsUtil.java
@@ -20,13 +20,10 @@
 package org.apache.pinot.segment.spi.index;
 
 import com.google.common.collect.Sets;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
-import org.apache.pinot.spi.config.table.FieldConfig;
 import org.apache.pinot.spi.config.table.IndexConfig;
 import org.apache.pinot.spi.config.table.TableConfig;
 import org.apache.pinot.spi.data.Schema;
@@ -105,24 +102,5 @@ public class FieldIndexConfigsUtil {
     public <C extends IndexConfig> ColumnConfigDeserializer<C> get(IndexType<C, ?, ?> indexType) {
       return indexType::getConfig;
     }
-  }
-
-  public static TableConfig createTableConfigFromOldFormat(TableConfig tableConfig, Schema schema) {
-    TableConfig clone = new TableConfig(tableConfig);
-    for (IndexType<?, ?, ?> indexType : IndexService.getInstance().getAllIndexes()) {
-      // get all the index data in new format
-      indexType.convertToNewFormat(clone, schema);
-    }
-    // cleanup the indexTypes field from all FieldConfig items
-    if (clone.getFieldConfigList() != null) {
-      List<FieldConfig> cleanFieldConfigList = new ArrayList<>();
-      for (FieldConfig fieldConfig : clone.getFieldConfigList()) {
-        cleanFieldConfigList.add(new FieldConfig.Builder(fieldConfig)
-            .withIndexTypes(null)
-            .build());
-      }
-      clone.setFieldConfigList(cleanFieldConfigList);
-    }
-    return clone;
   }
 }
